LayoutTests/fast/media/print-restores-previous-mediatype-expected.txt

Media type is set to 'handheld' to be different than the default 'screen'
PASS window.matchMedia('handheld').matches is true

The page is printed, which changes the media type to 'print', and then back
PASS window.matchMedia('handheld').matches is true
